pyproject.toml
[tool.poetry]
name = "short-url"
version = "0.1.0"
description = ""
authors = ["David Arias <chronos.software.dev@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
